Remove useless comments
This file was cluttered with absolutely useless comments that did nothing except make the file hard to read. This file is one of the few examples that shows how to use cors with firebase functions.

For example:
```js
// [START readBodyParam]
format = req.body.format;
 // [END readBodyParam]
```
This code snippet provides **zero** additional information, however it clutters the code.
To make matters worse, this exact case is repeated again, for the same code, with no change.

Comments should be used only for documentation, clarification, and nuances/things that need to be pointed out to future contributors.